fix columns in inep sinopse
