+++ models/br_inep_sinopse_estatistica_educacao_basica/br_inep_sinopse_estatistica_educacao_basica__docente_faixa_etaria_sexo.sql
@@ -15,7 +15,7 @@ select
     safe_cast(ano as int64) ano,
     safe_cast(sigla_uf as string) sigla_uf,
     safe_cast(id_municipio as string) id_municipio,
-    safe_cast(tipo_classe as string) tipo_classe,
+    safe_cast(etapa_ensino as string) etapa_ensino,
     safe_cast(faixa_etaria as string) faixa_etaria,
     safe_cast(sexo as string) sexo,
     safe_cast(quantidade_docente as int64) quantidade_docente,
