+++ models/br_inep_sinopse_estatistica_educacao_basica/br_inep_sinopse_estatistica_educacao_basica__docente_localizacao.sql
@@ -15,7 +15,7 @@ select
     safe_cast(ano as int64) ano,
     safe_cast(sigla_uf as string) sigla_uf,
     safe_cast(replace(id_municipio, ".0", "") as string) id_municipio,
-    safe_cast(tipo_classe as string) tipo_classe,
+    safe_cast(etapa_ensino as string) etapa_ensino,
     safe_cast(rede as string) rede,
     safe_cast(localizacao as string) localizacao,
     safe_cast(quantidade_docente as int64) quantidade_docente,
